finalize split-up between wep-api-testing adn web-api-documentation
