+++ src/hackingBuddyGPT/utils/web_api/custom_datatypes.py
